validaciones con zod a los endpoints agregadas

+++ src/controllers/ordersControllers.js
@@ -1,11 +1,9 @@
 const { Order } = require('../models/orderModel')
-
+const { orderSchema } = require('../common/orders/orders_schema');
+const { ObjectId } = require('mongodb');
 // Hacer un middleware para todas las rutas para verificar q el usuario esté autenticado
 
 const ordersGet = async (req, res) => {
-  // 1- Verificar q el usuario esté autenticado
-  // const token = req.cookies('nombre-de-la-cookie');
-  // console.log(token)
 
   const orders = await Order.find({})
 
@@ -18,13 +16,12 @@ const ordersGet = async (req, res) => {
 const orderGetById = async (req, res) => {
   const { id } = req.params
 
-  const order = await Order.findById(id) // buscar order por id y por el id del usuario
+  if(!id) return res.status(400).json({ message: 'ID es requerido' })
+  if(!ObjectId.isValid(id)) return res.status(400).json({ message: `ID --> ${id}, no es un id valido de mongodb ` })
 
-  if (!order)
-    return res.status(404).json({ message: `Orden con id _${id}_ no existe` })
+  const order = await Order.findById(id);
 
-  // verificar q la order exista
-  // verificar q la order pertenezca al usuario
+  if (!order) return res.status(404).json({ message: `Orden con id _${id}_ no existe` });
 
   res.status(200).json(order)
 }
@@ -32,18 +29,11 @@ const orderGetById = async (req, res) => {
 const orderPost = async (req, res) => {
   const order = req.body
 
-  const newOrder = new Order({
-    user: order.userId,
-    orderItems: order.orderItems,
-    shippingAddress: order.shippingAddress,
-    numberOfItems: order.numberOfItems,
-    subTotal: order.subTotal,
-    tax: order.tax,
-    total: order.total,
-    isPaid: order.isPaid,
-    paidAt: order.paidAt,
-    transactionId: order.transactionId
-  })
+  const check = orderSchema.safeParse(order);
+  if(!check.success) return res.status(400).json({ message: check.error.issues[0].message });
+
+
+  const newOrder = new Order(order);
 
   const data = await newOrder.save()
 
@@ -52,6 +42,9 @@ const orderPost = async (req, res) => {
 
 const orderDelete = async (req, res) => {
   const { id } = req.params
+
+  if(!id) return res.status(400).json({ message: 'ID es requerido' })
+  if(!ObjectId.isValid(id)) return res.status(400).json({ message: `ID --> ${id}, no es un id valido de mongodb ` })
 
   const order = await Order.findByIdAndDelete({ _id: id })
